frontalization2/scripts/requirements_eval.txt
lpips
facenet-pytorch
torchmetrics
torchvision
pytorch-fid
